Move the inline queue port guard into a shared fixture
The guard points every inline queue exchange this process makes at a port
nothing listens on, but it lived in the integration Harness, so the unit tests
had none: one that applies a staged patch settles it with the owner afterwards,
and that settle went to whatever tray happened to be running on the machine.

Now a module initializer in a file both test projects compile, so it runs before
any test in either and cannot be skipped by one that builds no harness. The port
is picked once per assembly rather than once per Harness: re-pointing per harness
would clobber a live InlineQueueHost if one were ever constructed first, which is
a worse risk than the OS reusing a port it just handed out.
