+++ src/Verify.Terminal.IntegrationTests/Verify.Terminal.IntegrationTests.csproj
@@ -12,4 +12,8 @@
     <PackageReference Include="xunit.v3" />
     <ProjectReference Include="..\Verify.Terminal\Verify.Terminal.csproj" />
   </ItemGroup>
+  <ItemGroup>
+    <!-- Shared with the other test project: both make inline queue exchanges. -->
+    <Compile Include="..\Shared\DeadInlineQueue.cs" Link="DeadInlineQueue.cs" />
+  </ItemGroup>
 </Project>
