+++ src/Verify.Terminal.Tests/GlobalUsings.cs
@@ -5,3 +5,4 @@ global using Shouldly;
 global using Spectre.Console.Testing;
 global using Spectre.IO;
 global using Spectre.IO.Testing;
+global using Verify.Terminal.Testing;
